+++ src/Quartz/Impl/AdoJobStore/JobStoreSupport.cs
@@ -322,20 +322,20 @@ namespace Quartz.Impl.AdoJobStore
 
         protected DbMetadata DbMetadata => ConnectionManager.GetDbMetadata(DataSource);
 
-        protected abstract ConnectionAndTransactionHolder GetNonManagedTXConnection();
+        protected abstract ValueTask<ConnectionAndTransactionHolder> GetNonManagedTXConnection();
 
         /// <summary>
         /// Gets the connection and starts a new transaction.
         /// </summary>
         /// <returns></returns>
-        protected virtual ConnectionAndTransactionHolder GetConnection()
+        protected virtual async ValueTask<ConnectionAndTransactionHolder> GetConnection()
         {
             DbConnection conn;
             DbTransaction tx;
             try
             {
                 conn = ConnectionManager.GetConnection(DataSource);
-                conn.Open();
+                await conn.OpenAsync();
             }
             catch (Exception e)
             {
@@ -3134,7 +3134,7 @@ namespace Quartz.Impl.AdoJobStore
             CancellationToken cancellationToken)
         {
             bool transOwner = false;
-            ConnectionAndTransactionHolder conn = GetNonManagedTXConnection();
+            ConnectionAndTransactionHolder conn = await GetNonManagedTXConnection();
             try
             {
                 RecoverMisfiredJobsResult result = RecoverMisfiredJobsResult.NoOp;
@@ -3206,7 +3206,7 @@ namespace Quartz.Impl.AdoJobStore
             bool transStateOwner = false;
             bool recovered = false;
 
-            ConnectionAndTransactionHolder conn = GetNonManagedTXConnection();
+            ConnectionAndTransactionHolder conn = await GetNonManagedTXConnection();
             try
             {
                 // Other than the first time, always checkin first to make sure there is
@@ -3947,7 +3947,7 @@ namespace Quartz.Impl.AdoJobStore
                     // until after acquiring the lock since it isn't needed.
                     if (LockHandler.RequiresConnection)
                     {
-                        conn = GetNonManagedTXConnection();
+                        conn = await GetNonManagedTXConnection();
                     }
 
                     transOwner = await LockHandler.ObtainLock(requestorId.Value, conn, lockName, cancellationToken).ConfigureAwait(false);
@@ -3955,7 +3955,7 @@ namespace Quartz.Impl.AdoJobStore
 
                 if (conn == null)
                 {
-                    conn = GetNonManagedTXConnection();
+                    conn = await GetNonManagedTXConnection();
                 }
 
                 T result = await txCallback(conn).ConfigureAwait(false);